Rename cparse to parse, and allow structure with dead code
