+++ src/frontegg-auth/src/client/user.rs
@@ -10,12 +10,12 @@
 use reqwest::Method;
 use serde::{Deserialize, Serialize};
 
-use crate::{client::CREATE_USERS_PATH, cparse::Paginated, error::Error};
+use crate::{client::CREATE_USERS_PATH, parse::Paginated, error::Error};
 
 use super::{Client, USERS_PATH};
 
 #[derive(Serialize)]
-pub struct NewUser {
+pub struct UserCreationRequest {
     email: String,
     name: String,
     provider: String,
@@ -23,8 +23,9 @@ pub struct NewUser {
     role_ids: Vec<String>,
 }
 
-#[derive(Deserialize, Debug)]
+#[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
+#[allow(dead_code)]
 pub struct User {
     id: String,
     name: String,
@@ -51,7 +52,7 @@ struct Role {
 
 #[derive(Debug, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
-pub struct CreateUser {
+pub struct UserCreationResponse {
     id: String,
     email: String,
     name: String,
@@ -83,13 +84,12 @@ impl Client {
     }
 
     /// Creates a new user in the authenticated organization.
-    pub async fn create_user(&self, new_user: NewUser) -> Result<CreateUser, Error> {
+    pub async fn create_user(&self, new_user: UserCreationRequest) -> Result<UserCreationResponse, Error> {
         let req = self.build_request(Method::POST, CREATE_USERS_PATH);
 
         let req = req.json(&new_user);
 
-        println!("Output: {:?}", req);
-        let new_user: CreateUser = self.send_request(req).await?;
-        Ok(new_user)
+        let user_created: UserCreationResponse = self.send_request(req).await?;
+        Ok(user_created)
     }
 }
